feat(extension): #1230: Implement onboarding screen with default frontend selection
