+++ apps/extension/src/routes/page/paths.ts
@@ -7,6 +7,7 @@ export enum PagePath {
   ONBOARDING_SUCCESS = '/welcome/success',
   SET_PASSWORD = '/welcome/set-password',
   SET_GRPC_ENDPOINT = '/welcome/set-grpc-endpoint',
+  SET_DEFAULT_FRONTEND = '/welcome/set-default-frontend',
   RESTORE_PASSWORD = '/restore-password',
   RESTORE_PASSWORD_INDEX = '/restore-password/',
   RESTORE_PASSWORD_SET_PASSWORD = '/restore-password/set-password',
